Move from string as errors to custom fail types
These change moves the crate from using strings as errors, which is a
pattern that should be used only for prototyping, to defining custom
fail types, based on https://boats.gitlab.io/failure/

I have decided to define four different error structs based on what is
written in this GitHub issue:
https://github.com/rust-lang-nursery/failure/issues/7

However, I am very open to change everything to a single error type
defined as an enum, if you prefer so.

In addition, this branch also resurrects a simple zigzag test that
slipped through our test revamp as commented.