Use a single &[u8] when decoding items in Reader
The current implementation creates a new view of `bytes` for each
iteration, meaning it is forced to read the first item in `bytes` over
and over, due to a
[gotcha](https://doc.rust-lang.org/src/std/io/impls.rs.html#180) in the
`io::Read` implementation of `&'a [u8]`.

The new implementation will consume bytes after each call to
`from_avro_datum`, effectively reading all the items.