Extend on the work by Kyle to support reference types
Instead of trying to resolve the references at parse time, we store references
in the AST, and then use the SchemaParseContext not only for parsing but also
for decoding. This allows us to support recursive type definitions.

I believe this patch should also include some improvements in how we handle
namespaces and aliases. Tests are still needed to prove aliases are working
correctly.

Tests are still needed for testing encoding.