Add schema compatibility to library
Most of this was transplanted from
https://github.com/apache/avro/blob/master/lang/ruby/lib/avro/schema_compatibility.rb as a
proof of concept.

What that means is that I'd be keen on early feedback on all things here, so that we can
get a good way to check for schema compatibility in this library, without me committing
too much (in terms of design) up front.