Port tests from apache/avro test_schema.py and implement schema equality based on parsing canonical form
I noticed that we were deriving PartialEq for Schema while my understanding
is that the spec asserts that equality between schemas has to rely on
parsing canonical form. So I manually impemented PartialEq and Eq based
on that.

I also created a bunch of issues to follow with the tests that I had to
comment because they were failing.