Fix union resolution
See comments in #96 for more info.

I am a bit scared of the performance implications in case of many
union fields with default value null to allow backward compatibility, so
I added an explicit branch in the default resolution to handle null
without a couple more allocations.

The most efficient solution would be to precompute the default value as
part of the schema parsing and store it as the default for the field
instead of its JSON representation, as we are instead doing now.